De-duplicating scenario data
Previously a scenario had all the default data, plus it's custom data. This made a very large and repetitive multimocks.js file

By merging default with selected on the client side we can signnificantly reduce the size of the mock data

This could potentially slow load of the client if there are a lot of requests to mock, however downloading a ~7MB file was also pretty slow!

This could be configurable if the previous behaviour is preferred.